Get child dependency for a package

Get child dependency for a package

child_dependency(package, fields = NULL)

Arguments

package
Package name.

fields
Which fields in DESCRIPTION? Values should be in Depends, Imports, LinkingTo, Suggests and Enhances.

Details

The dependency information is based on packages retrieved from CRAN/Bioconductor on 2021-10-28.

Value

A data frame with child packages as well as its heaviness on its child packages.

Examples

# \dontrun{
child_dependency("ComplexHeatmap")

#>           parents                  children dep_fields heaviness
#> 1  ComplexHeatmap                   airpart    Imports         7
#> 2  ComplexHeatmap                  AMARETTO    Depends         5
#> 3  ComplexHeatmap                     artMS   Suggests         9
#> 4  ComplexHeatmap                     bambu   Suggests        13
#> 5  ComplexHeatmap         BindingSiteFinder   Suggests        12
#> 6  ComplexHeatmap                BiocOncoTK    Imports        11
#> 7  ComplexHeatmap                   BioNERO    Imports         7
#> 8  ComplexHeatmap               blacksheepr    Imports        11
#> 9  ComplexHeatmap           BloodGen3Module    Imports         9
#> 10 ComplexHeatmap                BrainSABER   Suggests        15
#> 11 ComplexHeatmap                  CATALYST    Imports         6
#> 12 ComplexHeatmap                     celda    Imports         4
#> 13 ComplexHeatmap                      CeTF    Imports         9
#> 14 ComplexHeatmap                 clustifyr   Suggests        13
#> 15 ComplexHeatmap                 CNVRanger   Suggests        16
#> 16 ComplexHeatmap                     COCOA    Imports        12
#> 17 ComplexHeatmap                      cola    Imports         4
#> 18 ComplexHeatmap                cytoKernel    Imports        12
#> 19 ComplexHeatmap          DEComplexDisease    Imports        12
#> 20 ComplexHeatmap                 DEGreport    Imports         8
#> 21 ComplexHeatmap                       DEP    Imports         4
#> 22 ComplexHeatmap                   diffcyt    Imports         6
#> 23 ComplexHeatmap                   diffUTR    Imports        11
#> 24 ComplexHeatmap                  dittoSeq   Suggests        14
#> 25 ComplexHeatmap                     ELMER    Imports         3
#> 26 ComplexHeatmap           EnrichedHeatmap    Depends        11
#> 27 ComplexHeatmap         EnrichmentBrowser   Suggests        16
#> 28 ComplexHeatmap                     fCCAC    Imports        11
#> 29 ComplexHeatmap              gCrisprTools    Imports        13
#> 30 ComplexHeatmap                 GeneTonic    Imports         9
#> 31 ComplexHeatmap     GenomicSuperSignature    Imports        13
#> 32 ComplexHeatmap                    gmoviz    Imports         9
#> 33 ComplexHeatmap                  gtrellis   Suggests        12
#> 34 ComplexHeatmap              HilbertCurve   Suggests        14
#> 35 ComplexHeatmap InteractiveComplexHeatmap    Imports        12
#> 36 ComplexHeatmap               InterCellar    Imports         8
#> 37 ComplexHeatmap                      iSEE    Imports        11
#> 38 ComplexHeatmap              LineagePulse    Imports        10
#> 39 ComplexHeatmap               MatrixQCvis    Imports        12
#> 40 ComplexHeatmap                    MesKit    Imports        10
#> 41 ComplexHeatmap          microbiomeMarker    Imports         7
#> 42 ComplexHeatmap                      MOMA    Imports         9
#> 43 ComplexHeatmap                  monaLisa    Imports         5
#> 44 ComplexHeatmap                  msImpute   Suggests        14
#> 45 ComplexHeatmap                    muscat    Imports         8
#> 46 ComplexHeatmap                  musicatk    Imports         6
#> 47 ComplexHeatmap                 MWASTools    Imports        12
#> 48 ComplexHeatmap                 PathoStat    Imports         7
#> 49 ComplexHeatmap                   PeacoQC    Imports        10
#> 50 ComplexHeatmap                  pipeComp    Imports         6
#> 51 ComplexHeatmap                      POMA    Imports         6
#> 52 ComplexHeatmap               profileplyr    Imports         0
#> 53 ComplexHeatmap                  projectR   Suggests         8
#> 54 ComplexHeatmap                 QFeatures   Suggests        12
#> 55 ComplexHeatmap                    recoup    Depends         8
#> 56 ComplexHeatmap                     RLSeq    Imports         5
#> 57 ComplexHeatmap               scDblFinder   Suggests        12
#> 58 ComplexHeatmap                     sechm    Imports         9
#> 59 ComplexHeatmap                 segmenter    Imports        11
#> 60 ComplexHeatmap                   SEtools    Imports         5
#> 61 ComplexHeatmap        simplifyEnrichment    Imports         7
#> 62 ComplexHeatmap              singleCellTK    Imports         0
#> 63 ComplexHeatmap                   sparrow    Imports         9
#> 64 ComplexHeatmap                     spiky   Suggests        12
#> 65 ComplexHeatmap       TBSignatureProfiler    Imports        12
#> 66 ComplexHeatmap              TCGAbiolinks   Suggests        13
#> 67 ComplexHeatmap                 TCGAutils   Suggests        13
#> 68 ComplexHeatmap      TimeSeriesExperiment   Suggests        11
#> 69 ComplexHeatmap                   weitrix   Suggests        13
#> 70 ComplexHeatmap                      Xeva    Imports         8
#> 71 ComplexHeatmap                     YAPSA    Imports         2
#> 72 ComplexHeatmap            NanoporeRNASeq   Suggests        17
#> 73 ComplexHeatmap              TCGAWorkflow    Imports         0
#> 74 ComplexHeatmap                    armada    Imports        11
#> 75 ComplexHeatmap                  circlize   Suggests        20
#> 76 ComplexHeatmap                  CNVScope   Suggests         6
#> 77 ComplexHeatmap                     conos    Imports        18
#> 78 ComplexHeatmap               countToFPKM    Depends        19
#> 79 ComplexHeatmap                    eclust   Suggests         7
#> 80 ComplexHeatmap                    i2dash   Suggests        21
#> 81 ComplexHeatmap               IOHanalyzer   Suggests        19
#> 82 ComplexHeatmap                   MKomics    Imports        18
#> 83 ComplexHeatmap                      MOSS   Suggests        23
#> 84 ComplexHeatmap          multipanelfigure   Suggests        18
#> 85 ComplexHeatmap                   pkgndep    Imports        21
#> 86 ComplexHeatmap                   rKOMICS    Imports        13
#> 87 ComplexHeatmap                       RVA    Imports         8
#> 88 ComplexHeatmap                     sigQC    Imports         8
#> 89 ComplexHeatmap                 spiralize   Suggests        16
#> 90 ComplexHeatmap               tidyHeatmap    Imports        16
#> 91 ComplexHeatmap                visxhclust    Imports        15
#> 92 ComplexHeatmap                    wilson    Imports         4

# }