Get downstream dependency for a package

Get downstream dependency for a package

downstream_dependency(package, online = FALSE)

Arguments

package

Package name.

online

Whether use the newest package database directly from CRAN/Bioconductor or the pre-computated package database? The version of the pre-computated package database can be set via pkgndep_opt$heaviness_db_version.

Details

Downstream packages with relations of Depends, Imports and LinkingTo are retrieved.

Value

A data frame with all downstream packages.

Examples

# \dontrun{
downstream_dependency("ComplexHeatmap")
#>             parents                  children dep_fields heaviness
#> 18   ComplexHeatmap                  AMARETTO    Depends         5
#> 31   ComplexHeatmap           EnrichedHeatmap    Depends        11
#> 37   ComplexHeatmap InteractiveComplexHeatmap    Depends        12
#> 10   ComplexHeatmap               countToFPKM    Depends        19
#> 70   ComplexHeatmap                    recoup    Depends         7
#> 19   ComplexHeatmap                    ASURAT    Imports        13
#> 20   ComplexHeatmap                   BioNERO    Imports         7
#> 21   ComplexHeatmap                BiocOncoTK    Imports        10
#> 22   ComplexHeatmap           BloodGen3Module    Imports         9
#> 23   ComplexHeatmap                  CATALYST    Imports         7
#> 24   ComplexHeatmap                     COCOA    Imports        11
#> 25   ComplexHeatmap                     COTAN    Imports        16
#> 26   ComplexHeatmap                      CeTF    Imports         8
#> 27   ComplexHeatmap          DEComplexDisease    Imports        11
#> 28   ComplexHeatmap                 DEGreport    Imports         7
#> 29   ComplexHeatmap                       DEP    Imports         4
#> 30   ComplexHeatmap                     ELMER    Imports         3
#> 32   ComplexHeatmap                    FLAMES    Imports         6
#> 33   ComplexHeatmap                    GRaNIE    Imports         8
#> 34   ComplexHeatmap                 GeneTonic    Imports         8
#> 35   ComplexHeatmap     GenomicSuperSignature    Imports        13
#> 1    ComplexHeatmap                    IntLIM    Imports        14
#> 36   ComplexHeatmap               InterCellar    Imports         9
#> 38   ComplexHeatmap              LineagePulse    Imports        10
#> 2    ComplexHeatmap                   MKomics    Imports        18
#> 39   ComplexHeatmap                      MOMA    Imports         9
#> 40   ComplexHeatmap                 MWASTools    Imports        12
#> 41   ComplexHeatmap               MatrixQCvis    Imports        12
#> 42   ComplexHeatmap                    MesKit    Imports        10
#> 3    ComplexHeatmap                  MitoHEAR    Imports        10
#> 4    ComplexHeatmap                     PALMO    Imports         7
#> 43   ComplexHeatmap                      POMA    Imports         7
#> 44   ComplexHeatmap                 PathoStat    Imports         7
#> 45   ComplexHeatmap                   PeacoQC    Imports        12
#> 46   ComplexHeatmap                     RLSeq    Imports         5
#> 5    ComplexHeatmap                  RNAseqQC    Imports         7
#> 6    ComplexHeatmap                       RVA    Imports         8
#> 47   ComplexHeatmap                     SPIAT    Imports        10
#> 48   ComplexHeatmap                    SPONGE    Imports         9
#> 49   ComplexHeatmap       TBSignatureProfiler    Imports        11
#> 50   ComplexHeatmap                      Xeva    Imports         8
#> 51   ComplexHeatmap                     YAPSA    Imports         2
#> 52   ComplexHeatmap                   airpart    Imports         7
#> 7    ComplexHeatmap                    armada    Imports        12
#> 53   ComplexHeatmap               blacksheepr    Imports         8
#> 8    ComplexHeatmap              bulkAnalyseR    Imports         5
#> 54   ComplexHeatmap                     celda    Imports         4
#> 55   ComplexHeatmap                      cola    Imports         4
#> 9    ComplexHeatmap                     conos    Imports        19
#> 56   ComplexHeatmap                cytoKernel    Imports        12
#> 57   ComplexHeatmap                   diffUTR    Imports        10
#> 58   ComplexHeatmap                   diffcyt    Imports         8
#> 59   ComplexHeatmap                     fCCAC    Imports        10
#> 60   ComplexHeatmap              gCrisprTools    Imports        14
#> 61   ComplexHeatmap                    gmoviz    Imports         8
#> 62   ComplexHeatmap                    hermes    Imports         8
#> 63   ComplexHeatmap                      iSEE    Imports        11
#> 64   ComplexHeatmap          microbiomeMarker    Imports         6
#> 11   ComplexHeatmap                  missoNet    Imports        15
#> 65   ComplexHeatmap                  monaLisa    Imports         5
#> 66   ComplexHeatmap                    muscat    Imports         8
#> 67   ComplexHeatmap                  musicatk    Imports         6
#> 68   ComplexHeatmap                  pipeComp    Imports         6
#> 12   ComplexHeatmap                   pkgndep    Imports        21
#> 69   ComplexHeatmap               profileplyr    Imports         0
#> 13   ComplexHeatmap                   rKOMICS    Imports        14
#> 14   ComplexHeatmap                     scITD    Imports         4
#> 71   ComplexHeatmap                     sechm    Imports         9
#> 72   ComplexHeatmap                 segmenter    Imports        10
#> 73   ComplexHeatmap               signifinder    Imports         9
#> 74   ComplexHeatmap        simplifyEnrichment    Imports         7
#> 75   ComplexHeatmap              singleCellTK    Imports         0
#> 76   ComplexHeatmap                   sparrow    Imports         8
#> 15   ComplexHeatmap               tidyHeatmap    Imports        17
#> 16   ComplexHeatmap                visxhclust    Imports        15
#> 17   ComplexHeatmap                    wilson    Imports         4
#> 110 EnrichedHeatmap                extraChIPs    Imports        10
#> 210 EnrichedHeatmap               profileplyr    Imports         2
#> 111         BioNERO                 cageminer    Imports        42
#> 112       DEGreport                   isomiRs    Imports        24
#> 113           ELMER           TCGAbiolinksGUI    Imports        71
#> 114       MWASTools              MetaboSignal    Imports        41
#> 115          SPONGE                miRspongeR    Imports       132
#> 116           celda              singleCellTK    Imports        10
#> 211         diffcyt                   censcyt    Depends        53
#> 117         diffcyt                CyTOFpower    Imports        30
#> 310         diffcyt                   treekoR    Imports        55
#> 118            iSEE                   iSEEhex    Depends        33
#> 311            iSEE                     iSEEu    Depends         0
#> 212            iSEE                   iSEEhub    Imports        37
#> 119           sechm                   SEtools    Depends        19
#> 120      miRspongeR                     miRSM    Imports       149
#> 121         iSEEhex                     iSEEu    Depends         2
# }