Get functions that are imported to its child packages

Get functions that are imported to its child packages

get_all_functions_imported_to_children(package)

Arguments

package

Package name.

Details

The information is based on pre-computated results for a specific CRAN/Bioconductor snapshot. See pkgndep$heaviness_db_version for how to set the version of the snapshot.

Value

It returns a list of function names that are imported to every of its child packages.

Examples

# \dontrun{
get_all_functions_imported_to_children("circlize")
#> $MetCirc
#> [1] "circlize"               "circos.clear"           "circos.initialize"     
#> [4] "circos.link"            "circos.par"             "circos.text"           
#> [7] "circos.trackPlotRegion" "highlight.sector"      
#> 
#> $RNAsmc
#> [1] "circos.initialize"      "circos.trackPlotRegion" "circos.text"           
#> [4] "circos.link"            "circos.clear"          
#> 
#> $countToFPKM
#> [1] "colorRamp2"
#> 
#> $yarrr
#> [1] "colorRamp2"
#> 
#> $AMARETTO
#> [1] "colorRamp2" "rand_color"
#> 
#> $CATALYST
#> [1] "colorRamp2"
#> 
#> $COTAN
#> [1] "colorRamp2"
#> 
#> $CeTF
#>  [1] "CELL_META"            "circos.clear"         "circos.initialize"   
#>  [4] "circos.link"          "circos.par"           "circos.rect"         
#>  [7] "circos.text"          "circos.track"         "colorRamp2"          
#> [10] "highlight.chromosome" "rand_color"          
#> 
#> $ComplexHeatmap
#> [1] "add_transparency" "colorRamp2"       "rand_color"       "smartAlign"      
#> 
#> $DEGreport
#> [1] "colorRamp2"
#> 
#> $DEsubs
#> [1] "circos.par"           "chordDiagram"         "get.all.sector.index"
#> [4] "get.cell.meta.data"   "circos.text"          "circos.clear"        
#> 
#> $EnrichedHeatmap
#> [1] "colorRamp2"
#> 
#> $FLAMES
#> [1] "colorRamp2"
#> 
#> $GALLO
#> [1] "chordDiagram"       "circos.clear"       "circos.text"       
#> [4] "circos.track"       "get.cell.meta.data"
#> 
#> $GRaNIE
#> [1] "colorRamp2"
#> 
#> $GeneTonic
#> [1] "colorRamp2"
#> 
#> $HilbertCurve
#> [1] "rand_color"     "read.chromInfo"
#> 
#> $InterCellar
#> [1] "CELL_META"              "chordDiagram"           "circos.clear"          
#> [4] "circos.par"             "circos.text"            "circos.trackPlotRegion"
#> [7] "get.cell.meta.data"     "highlight.sector"       "uh"                    
#> 
#> $LymphoSeq
#> [1] "chordDiagram" "colorRamp2"  
#> 
#> $MKomics
#> [1] "colorRamp2"
#> 
#> $MoonlightR
#> [1] "circlize"               "circos.clear"           "circos.initialize"     
#> [4] "circos.par"             "circos.rect"            "circos.text"           
#> [7] "circos.trackPlotRegion" "get.cell.meta.data"    
#> 
#> $PeacoQC
#> [1] "colorRamp2"
#> 
#> $RVA
#> [1] "colorRamp2"
#> 
#> $RepertoiR
#> [1] "circos.barplot"     "circos.clear"       "circos.initialize" 
#> [4] "circos.link"        "circos.par"         "circos.rect"       
#> [7] "circos.track"       "get.cell.meta.data" "mm_h"              
#> 
#> $SEtools
#> [1] "colorRamp2"
#> 
#> $SingleCellSignalR
#> [1] "chordDiagramFromDataFrame" "circos.axis"              
#> [3] "circos.text"               "circos.trackPlotRegion"   
#> [5] "get.cell.meta.data"       
#> 
#> $TRONCO
#> [1] "chordDiagram"           "circos.clear"           "circos.par"            
#> [4] "circos.text"            "circos.trackPlotRegion" "get.cell.meta.data"    
#> 
#> $YAPSA
#> [1] "rainfallTransform"
#> 
#> $acc
#> [1] "circos.clear"           "circos.initialize"      "circos.lines"          
#> [4] "circos.par"             "circos.rect"            "circos.text"           
#> [7] "circos.trackPlotRegion" "get.cell.meta.data"    
#> 
#> $bandle
#> [1] "chordDiagram"           "circos.trackPlotRegion" "CELL_META"             
#> [4] "get.cell.meta.data"     "circos.text"            "circos.axis"           
#> [7] "circos.par"             "circos.clear"           "cm_h"                  
#> 
#> $capm
#> [1] "chordDiagram"           "circos.axis"            "circos.clear"          
#> [4] "circos.par"             "circos.text"            "circos.trackPlotRegion"
#> [7] "get.cell.meta.data"    
#> 
#> $comapr
#> [1] "colorRamp2"
#> 
#> $cytoKernel
#> [1] "colorRamp2"
#> 
#> $diffcyt
#> [1] "colorRamp2"
#> 
#> $ethnobotanyR
#> [1] "chordDiagram"       "circos.text"        "get.cell.meta.data"
#> 
#> $gtrellis
#> [1] "cytoband.col"   "read.chromInfo" "read.cytoband" 
#> 
#> $hermes
#> [1] "add_transparency"
#> 
#> $iSEE
#> [1] "colorRamp2"
#> 
#> $immunarch
#> [1] "chordDiagram"
#> 
#> $monaLisa
#> [1] "colorRamp2"
#> 
#> $pipeComp
#> [1] "colorRamp2"
#> 
#> $profileplyr
#> [1] "colorRamp2"
#> 
#> $rKOMICS
#> [1] "colorRamp2"
#> 
#> $recoup
#> [1] "colorRamp2"
#> 
#> $rexposome
#> [1] "circos.clear"           "circos.initialize"      "circos.link"           
#> [4] "circos.text"            "circos.trackPlotRegion" "get.cell.meta.data"    
#> 
#> $scITD
#> [1] "colorRamp2"
#> 
#> $sechm
#> [1] "colorRamp2"
#> 
#> $simplifyEnrichment
#> [1] "colorRamp2" "rand_color"
#> 
#> $socialh
#> [1] "chordDiagram" "circos.clear" "circos.par"   "circos.text"  "circos.track"
#> 
#> $sparrow
#> [1] "colorRamp2"
#> 
#> $spiralize
#> [1] "colorRamp2" "rand_color"
#> 
#> $tidyHeatmap
#> [1] "colorRamp2"
#> 
#> $visxhclust
#> [1] "colorRamp2"
#> 
# }