Load dependency data of all packages

Load dependency data of all packages

load_all_pkg_dep(hash = TRUE)

Arguments

hash
Whether to convert the named list to a hash table by hash.

Details

It loads the package dependency analysis for all CRAN/Biocondutor packages done on 2021-10-28.

Value

A list (as a hash table) of pkgndep objects where each element corresponds to the analysis on one package.

Examples

# \dontrun{
lt = load_all_pkg_dep()
length(lt)

#> [1] 21741

head(names(lt))

#> [1] "A3"            "AATtools"      "ABACUS"        "ABAData"      
#> [5] "ABAEnrichment" "ABC.RAP"      

lt[["ggplot2"]]

#> ggplot2, version 3.3.5
#> - 37 additional packages are required for installing 'ggplot2'.
#> - 137 additional packages are required if installing packages listed
#>   in all fields in DESCRIPTION.

# }