Load package database

Load package database

load_pkg_db(lib = NULL, online = TRUE, db = NULL, verbose = TRUE)

Arguments

lib

Local library path. If the value is NA, only remote package database is used.

online

If the value is TRUE, it will directly use the newest package database file from CRAN/Bioconductor. If the value is FALSE, it uses the pre-computated package database on a specific CRAN/Bioconductor snapshot. The version of the pre-computated package database can be set via pkgndep_opt$heaviness_db_version.

db

A pre-computed pkg_db object.

verbose

Whetehr to print messages.

Details

It loads the package database from CRAN/Bioconductor and locally installed packages.

The database object internaly is cached for repeated use of other functions in this package.

Value

A pkg_db class object. See reformat_db for how to use the pkg_db object.

Examples

# \dontrun{
pkg_db = load_pkg_db(lib = NA)
#> retrieve package database from CRAN/Bioconductor (3.16)...
#>   - 22451 remote packages on CRAN/Bioconductor.
#> prepare dependency table...
#> prepare reverse dependency table...
pkg_db
#> A package database of 22451 packages.
#>   - 0 local / 22451 remote packages.
#>   - 18961 CRAN / 3490 Bioconductor / 0 other packages.
# }