Load all package dependency statistics

Load all package dependency statistics

load_pkg_stat_snapshot()

Details

It loads the package dependency analysis for all CRAN/Biocondutor packages done on 2021-10-28.

Value

A data frame of various columns.

Examples

# \dontrun{
df = load_pkg_stat_snapshot()
head(df)
#>                 package                                              repository
#> a4                   a4 https://bioconductor.org/packages/3.15/bioc/src/contrib
#> a4Base           a4Base https://bioconductor.org/packages/3.15/bioc/src/contrib
#> a4Classif     a4Classif https://bioconductor.org/packages/3.15/bioc/src/contrib
#> a4Core           a4Core https://bioconductor.org/packages/3.15/bioc/src/contrib
#> a4Preproc     a4Preproc https://bioconductor.org/packages/3.15/bioc/src/contrib
#> a4Reporting a4Reporting https://bioconductor.org/packages/3.15/bioc/src/contrib
#>             n_by_strong n_by_all n_parents max_rel_heaviness_from_parents
#> a4                   82      162         5                       2.044444
#> a4Base               73      154        13                       1.050633
#> a4Classif            32       89        11                       1.166667
#> a4Core               19       43         4                       1.812500
#> a4Preproc             6       66         2                       1.066667
#> a4Reporting           4       29         2                       1.272727
#>             max_heaviness_from_parents adjusted_max_heaviness_from_parents
#> a4                                  47                          20.0609756
#> a4Base                               4                           2.0975610
#> a4Classif                            6                           3.0000000
#> a4Core                              13                           5.3902439
#> a4Preproc                            1                           0.3902439
#> a4Reporting                          3                           1.1707317
#>             heaviest_parent
#> a4                   a4Base
#> a4Base           genefilter
#> a4Classif              ROCR
#> a4Core               glmnet
#> a4Preproc           Biobase
#> a4Reporting          xtable
#>                                                                                                                                                                                                      max_heaviness_parent_info
#> a4                  Parent package &lsquo;a4Base&rsquo; contibutes the highest heaviness on &lsquo;a4&rsquo;. It is listed in &lsquo;Depends&rsquo but no object from parent is imported to the namespace of &lsquo;a4&rsquo;.
#> a4Base                                            Parent package &lsquo;genefilter&rsquo; contibutes the highest heaviness on &lsquo;a4Base&rsquo;. 5 functions/objects are imported to the namespace of &lsquo;a4Base&rsquo;.
#> a4Classif                                         Parent package &lsquo;ROCR&rsquo; contibutes the highest heaviness on &lsquo;a4Classif&rsquo;. 3 functions/objects are imported to the namespace of &lsquo;a4Classif&rsquo;.
#> a4Core      Parent package &lsquo;glmnet&rsquo; contibutes the highest heaviness on &lsquo;a4Core&rsquo;. The whole set of functions/methods/classes from parent package is imported to the namespace of &lsquo;a4Core&rsquo;.
#> a4Preproc                                      Parent package &lsquo;Biobase&rsquo; contibutes the highest heaviness on &lsquo;a4Preproc&rsquo;. 7 functions/objects are imported to the namespace of &lsquo;a4Preproc&rsquo;.
#> a4Reporting                                 Parent package &lsquo;xtable&rsquo; contibutes the highest heaviness on &lsquo;a4Reporting&rsquo;. 1 functions/objects are imported to the namespace of &lsquo;a4Reporting&rsquo;.
#>             total_heaviness_from_parents adjusted_total_heaviness_from_parents
#> a4                                    55                            30.3686697
#> a4Base                                15                             9.5157123
#> a4Classif                             13                             7.9931373
#> a4Core                                15                             8.1150267
#> a4Preproc                              1                             0.5179698
#> a4Reporting                            4                             2.0718791
#>             reducible max_co_heaviness_from_parents
#> a4              FALSE                             5
#> a4Base           TRUE                            35
#> a4Classif        TRUE                             7
#> a4Core          FALSE                             2
#> a4Preproc        TRUE                             5
#> a4Reporting      TRUE                             0
#>             max_co_heaviness_parents_pair max_co_heaviness_parents_pair_type
#> a4                       a4Base,a4Classif               have-common-upstream
#> a4Base                 annaffy,genefilter               have-common-upstream
#> a4Classif                   a4Core,glmnet                       parent-child
#> a4Core                     Biobase,glmnet                                   
#> a4Preproc            Biobase,BiocGenerics                       parent-child
#> a4Reporting                                                                 
#>             heaviness_on_children n_children heaviness_on_downstream
#> a4                      0.0000000          0                    0.00
#> a4Base                 47.0000000          1                   47.00
#> a4Classif               7.0000000          1                    7.00
#> a4Core                  2.0000000          4                    2.25
#> a4Preproc               0.6666667          3                    1.00
#> a4Reporting             1.0000000          1                    1.00
#>             n_downstream hv_downstream_values adjusted_heaviness_on_children
#> a4                     0                                          0.00000000
#> a4Base                 1                   47                     4.27272727
#> a4Classif              1                    7                     0.63636364
#> a4Core                 4           1, 1, 1, 6                     0.57142857
#> a4Preproc              3              1, 1, 1                     0.15384615
#> a4Reporting            1                    1                     0.09090909
#>             adjusted_heaviness_on_downstream heaviness_on_indirect_downstream
#> a4                                 0.0000000                                0
#> a4Base                             2.9375000                                0
#> a4Classif                          0.4375000                                0
#> a4Core                             0.4736842                                0
#> a4Preproc                          0.1666667                                0
#> a4Reporting                        0.0625000                                0
#>             n_indirect_downstream hv_indirect_downstream_values
#> a4                              0                              
#> a4Base                          0                              
#> a4Classif                       0                              
#> a4Core                          0                              
#> a4Preproc                       0                              
#> a4Reporting                     0                              
#>             adjusted_heaviness_on_indirect_downstream
#> a4                                                  0
#> a4Base                                              0
#> a4Classif                                           0
#> a4Core                                              0
#> a4Preproc                                           0
#> a4Reporting                                         0
# }