Loaded packages

Loaded packages

loaded_packages(pkg, verbose = TRUE)

Arguments

pkg
A package name.

verbose
Whether to print messages.

Details

It loads package pkg in a new R session and collects which other packages are loaded from sessionInfo.

Value

A data frame.

Examples

loaded_packages("ComplexHeatmap")

#> Loading ComplexHeatmap to a new R session... 31 namespaces loaded.

#>              pkg       type loading_time
#> 1           grid   basePkgs        1.262
#> 2          stats   basePkgs        1.262
#> 3       graphics   basePkgs        1.262
#> 4      grDevices   basePkgs        1.262
#> 5          utils   basePkgs        1.262
#> 6       datasets   basePkgs        1.262
#> 7        methods   basePkgs        1.262
#> 8           base   basePkgs        1.262
#> 9       circlize loadedOnly        1.262
#> 10   matrixStats loadedOnly        1.262
#> 11     codetools loadedOnly        1.262
#> 12       IRanges loadedOnly        1.262
#> 13           png loadedOnly        1.262
#> 14        digest loadedOnly        1.262
#> 15       foreach loadedOnly        1.262
#> 16        crayon loadedOnly        1.262
#> 17        stats4 loadedOnly        1.262
#> 18 GlobalOptions loadedOnly        1.262
#> 19    doParallel loadedOnly        1.262
#> 20     S4Vectors loadedOnly        1.262
#> 21    GetoptLong loadedOnly        1.262
#> 22  RColorBrewer loadedOnly        1.262
#> 23         rjson loadedOnly        1.262
#> 24     iterators loadedOnly        1.262
#> 25      parallel loadedOnly        1.262
#> 26      compiler loadedOnly        1.262
#> 27          clue loadedOnly        1.262
#> 28    colorspace loadedOnly        1.262
#> 29  BiocGenerics loadedOnly        1.262
#> 30       cluster loadedOnly        1.262
#> 31         shape loadedOnly        1.262