Get parent dependency for a package

Get parent dependency for a package

parent_dependency(package, fields = NULL)

Arguments

package
Package name.

fields
Which fields in DESCRIPTION? Values should be in Depends, Imports, LinkingTo, Suggests and Enhances.

Details

The dependency information is based on packages retrieved from CRAN/Bioconductor on 2021-10-28.

Value

A data frame with parent packages as well as their heaviness on pacakge.

Examples

parent_dependency("ComplexHeatmap")

#>            parents       children dep_fields heaviness
#> 1          methods ComplexHeatmap    Depends         0
#> 2             grid ComplexHeatmap    Depends         0
#> 3         graphics ComplexHeatmap    Depends         0
#> 4            stats ComplexHeatmap    Depends         0
#> 5        grDevices ComplexHeatmap    Depends         0
#> 6         circlize ComplexHeatmap    Imports         2
#> 7       GetoptLong ComplexHeatmap    Imports         3
#> 8       colorspace ComplexHeatmap    Imports         0
#> 9             clue ComplexHeatmap    Imports         2
#> 10    RColorBrewer ComplexHeatmap    Imports         1
#> 11   GlobalOptions ComplexHeatmap    Imports         0
#> 12             png ComplexHeatmap    Imports         1
#> 13          digest ComplexHeatmap    Imports         1
#> 14         IRanges ComplexHeatmap    Imports         4
#> 15     matrixStats ComplexHeatmap    Imports         1
#> 16         foreach ComplexHeatmap    Imports         0
#> 17      doParallel ComplexHeatmap    Imports         2
#> 18        testthat ComplexHeatmap   Suggests        30
#> 19           knitr ComplexHeatmap   Suggests        10
#> 20        markdown ComplexHeatmap   Suggests         4
#> 21        dendsort ComplexHeatmap   Suggests         1
#> 22            jpeg ComplexHeatmap   Suggests         1
#> 23            tiff ComplexHeatmap   Suggests         1
#> 24     fastcluster ComplexHeatmap   Suggests         1
#> 25 EnrichedHeatmap ComplexHeatmap   Suggests        13
#> 26      dendextend ComplexHeatmap   Suggests        32
#> 27        grImport ComplexHeatmap   Suggests         2
#> 28       grImport2 ComplexHeatmap   Suggests         4
#> 29            glue ComplexHeatmap   Suggests         1
#> 30   GenomicRanges ComplexHeatmap   Suggests         8
#> 31        gridtext ComplexHeatmap   Suggests        15
#> 32        pheatmap ComplexHeatmap   Suggests        11
#> 33    gridGraphics ComplexHeatmap   Suggests         1
#> 34          gplots ComplexHeatmap   Suggests         5
#> 35       rmarkdown ComplexHeatmap   Suggests        18
#> 36           Cairo ComplexHeatmap   Suggests         1