Make the dependency heatmap

Make the dependency heatmap

# S3 method for pkgndep
plot(x, pkg_fontsize = 10*cex, title_fontsize = 12*cex,
    legend_fontsize = 10*cex, fix_size = !dev.interactive(), cex = 1,
    help = TRUE, file = NULL, res = 144, ...)

Arguments

x

An object from pkgndep.

pkg_fontsize

Font size for the package names.

title_fontsize

Font size for the title.

legend_fontsize

Font size for the legends.

fix_size

Should the rows and columns in the heatmap have fixed size?

cex

A factor multiplicated to all font sizes.

help

Whether to print help message?

file

A path of the figure. The size of the figure is automatically calculated.

res

Resolution of the figure (only for png and jpeg).

...

Other arguments.

Details

If fix_size is set to TRUE. The size of the whole plot can be obtained by:

size = plot(x, fix_size = TRUE)

where size is a numeric vector of length two which are the width and height of the whole heatmap. If file argument is set, the size of the figure is automatically calculated.

If there are no dependency packages stored in x, NULL is returned.

Value

A vector of two numeric values (in inches) that correspond to the width and height of the plot.

Examples

# See examples in `pkgndep()`.