Format the package database

Format the package database

reformat_db(db)

Arguments

db

A data frame returned from available.packages or installed.packages.

Details

It reformats the data frame of the package database into a pkg_db class object.

Value

A pkg_db class object. There are the following methods:

pkg_db$get_meta(package,field=NULL)

field can take values in "Package", "Version" and "Repository".

pkg_db$get_dependency_table(package)

Get the dependency table.

pkg_db$get_rev_dependency_table(package)

Get the reverse dependency table.

pkg_db$package_dependencies(package,recursive=FALSE,reverse=FALSE,which="strong",simplify=FALSE)

All the arguments are the same as in package_dependencies. Argument simplify controls whether to return a data frame or a simplied vector.

Examples

# \dontrun{
db = available.packages()
db2 = reformat_db(db)
#> prepare dependency table...
#> prepare reverse dependency table...

# a pkg_db object generated on 2021-10-28 can be loaded by load_pkg_db()
db2 = load_pkg_db(snapshot = TRUE)
db2
#> A package database of 22076 packages.
#>   - 0 local / 22076 remote packages.
#>   - 3438 Bioconductor / 18638 CRAN / 0 other packages.
db2$get_meta("ComplexHeatmap")
#>                       Package Version Local
#> ComplexHeatmap ComplexHeatmap  2.12.0    no
#>                                                             Repository
#> ComplexHeatmap https://bioconductor.org/packages/3.15/bioc/src/contrib
db2$get_dependency_table("ComplexHeatmap")
#>       package          dependency        dep_fields
#>  [1,] "ComplexHeatmap" "methods"         "Depends" 
#>  [2,] "ComplexHeatmap" "grid"            "Depends" 
#>  [3,] "ComplexHeatmap" "graphics"        "Depends" 
#>  [4,] "ComplexHeatmap" "stats"           "Depends" 
#>  [5,] "ComplexHeatmap" "grDevices"       "Depends" 
#>  [6,] "ComplexHeatmap" "circlize"        "Imports" 
#>  [7,] "ComplexHeatmap" "GetoptLong"      "Imports" 
#>  [8,] "ComplexHeatmap" "colorspace"      "Imports" 
#>  [9,] "ComplexHeatmap" "clue"            "Imports" 
#> [10,] "ComplexHeatmap" "RColorBrewer"    "Imports" 
#> [11,] "ComplexHeatmap" "GlobalOptions"   "Imports" 
#> [12,] "ComplexHeatmap" "png"             "Imports" 
#> [13,] "ComplexHeatmap" "digest"          "Imports" 
#> [14,] "ComplexHeatmap" "IRanges"         "Imports" 
#> [15,] "ComplexHeatmap" "matrixStats"     "Imports" 
#> [16,] "ComplexHeatmap" "foreach"         "Imports" 
#> [17,] "ComplexHeatmap" "doParallel"      "Imports" 
#> [18,] "ComplexHeatmap" "codetools"       "Imports" 
#> [19,] "ComplexHeatmap" "testthat"        "Suggests"
#> [20,] "ComplexHeatmap" "knitr"           "Suggests"
#> [21,] "ComplexHeatmap" "markdown"        "Suggests"
#> [22,] "ComplexHeatmap" "dendsort"        "Suggests"
#> [23,] "ComplexHeatmap" "jpeg"            "Suggests"
#> [24,] "ComplexHeatmap" "tiff"            "Suggests"
#> [25,] "ComplexHeatmap" "fastcluster"     "Suggests"
#> [26,] "ComplexHeatmap" "EnrichedHeatmap" "Suggests"
#> [27,] "ComplexHeatmap" "dendextend"      "Suggests"
#> [28,] "ComplexHeatmap" "grImport"        "Suggests"
#> [29,] "ComplexHeatmap" "grImport2"       "Suggests"
#> [30,] "ComplexHeatmap" "glue"            "Suggests"
#> [31,] "ComplexHeatmap" "GenomicRanges"   "Suggests"
#> [32,] "ComplexHeatmap" "gridtext"        "Suggests"
#> [33,] "ComplexHeatmap" "pheatmap"        "Suggests"
#> [34,] "ComplexHeatmap" "gridGraphics"    "Suggests"
#> [35,] "ComplexHeatmap" "gplots"          "Suggests"
#> [36,] "ComplexHeatmap" "rmarkdown"       "Suggests"
#> [37,] "ComplexHeatmap" "Cairo"           "Suggests"
db2$get_rev_dependency_table("ComplexHeatmap")
#>        package                     dependency       dep_fields
#>   [1,] "airpart"                   "ComplexHeatmap" "Imports" 
#>   [2,] "AMARETTO"                  "ComplexHeatmap" "Depends" 
#>   [3,] "artMS"                     "ComplexHeatmap" "Suggests"
#>   [4,] "ASURAT"                    "ComplexHeatmap" "Imports" 
#>   [5,] "bambu"                     "ComplexHeatmap" "Suggests"
#>   [6,] "BindingSiteFinder"         "ComplexHeatmap" "Suggests"
#>   [7,] "BiocOncoTK"                "ComplexHeatmap" "Imports" 
#>   [8,] "BioNERO"                   "ComplexHeatmap" "Imports" 
#>   [9,] "blacksheepr"               "ComplexHeatmap" "Imports" 
#>  [10,] "BloodGen3Module"           "ComplexHeatmap" "Imports" 
#>  [11,] "BrainSABER"                "ComplexHeatmap" "Suggests"
#>  [12,] "CATALYST"                  "ComplexHeatmap" "Imports" 
#>  [13,] "celda"                     "ComplexHeatmap" "Imports" 
#>  [14,] "CeTF"                      "ComplexHeatmap" "Imports" 
#>  [15,] "clustifyr"                 "ComplexHeatmap" "Suggests"
#>  [16,] "CNVRanger"                 "ComplexHeatmap" "Suggests"
#>  [17,] "COCOA"                     "ComplexHeatmap" "Imports" 
#>  [18,] "cola"                      "ComplexHeatmap" "Imports" 
#>  [19,] "COTAN"                     "ComplexHeatmap" "Imports" 
#>  [20,] "cytoKernel"                "ComplexHeatmap" "Imports" 
#>  [21,] "DEComplexDisease"          "ComplexHeatmap" "Imports" 
#>  [22,] "DEGreport"                 "ComplexHeatmap" "Imports" 
#>  [23,] "DEP"                       "ComplexHeatmap" "Imports" 
#>  [24,] "diffcyt"                   "ComplexHeatmap" "Imports" 
#>  [25,] "diffUTR"                   "ComplexHeatmap" "Imports" 
#>  [26,] "dittoSeq"                  "ComplexHeatmap" "Suggests"
#>  [27,] "ELMER"                     "ComplexHeatmap" "Imports" 
#>  [28,] "EnrichedHeatmap"           "ComplexHeatmap" "Depends" 
#>  [29,] "EnrichmentBrowser"         "ComplexHeatmap" "Suggests"
#>  [30,] "fCCAC"                     "ComplexHeatmap" "Imports" 
#>  [31,] "FLAMES"                    "ComplexHeatmap" "Imports" 
#>  [32,] "gCrisprTools"              "ComplexHeatmap" "Imports" 
#>  [33,] "GeneTonic"                 "ComplexHeatmap" "Imports" 
#>  [34,] "GenomicSuperSignature"     "ComplexHeatmap" "Imports" 
#>  [35,] "gmoviz"                    "ComplexHeatmap" "Imports" 
#>  [36,] "GRaNIE"                    "ComplexHeatmap" "Imports" 
#>  [37,] "gtrellis"                  "ComplexHeatmap" "Suggests"
#>  [38,] "hermes"                    "ComplexHeatmap" "Imports" 
#>  [39,] "HilbertCurve"              "ComplexHeatmap" "Suggests"
#>  [40,] "InteractiveComplexHeatmap" "ComplexHeatmap" "Depends" 
#>  [41,] "InterCellar"               "ComplexHeatmap" "Imports" 
#>  [42,] "iSEE"                      "ComplexHeatmap" "Imports" 
#>  [43,] "LineagePulse"              "ComplexHeatmap" "Imports" 
#>  [44,] "MatrixQCvis"               "ComplexHeatmap" "Imports" 
#>  [45,] "MesKit"                    "ComplexHeatmap" "Imports" 
#>  [46,] "microbiomeMarker"          "ComplexHeatmap" "Imports" 
#>  [47,] "MOMA"                      "ComplexHeatmap" "Imports" 
#>  [48,] "monaLisa"                  "ComplexHeatmap" "Imports" 
#>  [49,] "msImpute"                  "ComplexHeatmap" "Suggests"
#>  [50,] "muscat"                    "ComplexHeatmap" "Imports" 
#>  [51,] "musicatk"                  "ComplexHeatmap" "Imports" 
#>  [52,] "MWASTools"                 "ComplexHeatmap" "Imports" 
#>  [53,] "pareg"                     "ComplexHeatmap" "Suggests"
#>  [54,] "PathoStat"                 "ComplexHeatmap" "Imports" 
#>  [55,] "PeacoQC"                   "ComplexHeatmap" "Imports" 
#>  [56,] "pipeComp"                  "ComplexHeatmap" "Imports" 
#>  [57,] "plotgardener"              "ComplexHeatmap" "Suggests"
#>  [58,] "POMA"                      "ComplexHeatmap" "Imports" 
#>  [59,] "profileplyr"               "ComplexHeatmap" "Imports" 
#>  [60,] "projectR"                  "ComplexHeatmap" "Suggests"
#>  [61,] "QFeatures"                 "ComplexHeatmap" "Suggests"
#>  [62,] "recoup"                    "ComplexHeatmap" "Depends" 
#>  [63,] "RLSeq"                     "ComplexHeatmap" "Imports" 
#>  [64,] "scDblFinder"               "ComplexHeatmap" "Suggests"
#>  [65,] "sechm"                     "ComplexHeatmap" "Imports" 
#>  [66,] "segmenter"                 "ComplexHeatmap" "Imports" 
#>  [67,] "simplifyEnrichment"        "ComplexHeatmap" "Imports" 
#>  [68,] "singleCellTK"              "ComplexHeatmap" "Imports" 
#>  [69,] "sparrow"                   "ComplexHeatmap" "Imports" 
#>  [70,] "spiky"                     "ComplexHeatmap" "Suggests"
#>  [71,] "SPONGE"                    "ComplexHeatmap" "Imports" 
#>  [72,] "TBSignatureProfiler"       "ComplexHeatmap" "Imports" 
#>  [73,] "TCGAbiolinks"              "ComplexHeatmap" "Suggests"
#>  [74,] "TCGAutils"                 "ComplexHeatmap" "Suggests"
#>  [75,] "TimeSeriesExperiment"      "ComplexHeatmap" "Suggests"
#>  [76,] "weitrix"                   "ComplexHeatmap" "Suggests"
#>  [77,] "Xeva"                      "ComplexHeatmap" "Imports" 
#>  [78,] "YAPSA"                     "ComplexHeatmap" "Imports" 
#>  [79,] "NanoporeRNASeq"            "ComplexHeatmap" "Suggests"
#>  [80,] "armada"                    "ComplexHeatmap" "Imports" 
#>  [81,] "bulkAnalyseR"              "ComplexHeatmap" "Imports" 
#>  [82,] "CIARA"                     "ComplexHeatmap" "Suggests"
#>  [83,] "circlize"                  "ComplexHeatmap" "Suggests"
#>  [84,] "CNVScope"                  "ComplexHeatmap" "Suggests"
#>  [85,] "conos"                     "ComplexHeatmap" "Imports" 
#>  [86,] "countToFPKM"               "ComplexHeatmap" "Depends" 
#>  [87,] "eclust"                    "ComplexHeatmap" "Suggests"
#>  [88,] "i2dash"                    "ComplexHeatmap" "Suggests"
#>  [89,] "IOHanalyzer"               "ComplexHeatmap" "Suggests"
#>  [90,] "MitoHEAR"                  "ComplexHeatmap" "Imports" 
#>  [91,] "MKomics"                   "ComplexHeatmap" "Imports" 
#>  [92,] "MOSS"                      "ComplexHeatmap" "Suggests"
#>  [93,] "multipanelfigure"          "ComplexHeatmap" "Suggests"
#>  [94,] "PALMO"                     "ComplexHeatmap" "Imports" 
#>  [95,] "pkgndep"                   "ComplexHeatmap" "Imports" 
#>  [96,] "rKOMICS"                   "ComplexHeatmap" "Imports" 
#>  [97,] "RNAseqQC"                  "ComplexHeatmap" "Imports" 
#>  [98,] "RVA"                       "ComplexHeatmap" "Imports" 
#>  [99,] "scITD"                     "ComplexHeatmap" "Imports" 
#> [100,] "spiralize"                 "ComplexHeatmap" "Suggests"
#> [101,] "tidyHeatmap"               "ComplexHeatmap" "Imports" 
#> [102,] "tinyarray"                 "ComplexHeatmap" "Suggests"
#> [103,] "visxhclust"                "ComplexHeatmap" "Imports" 
#> [104,] "wilson"                    "ComplexHeatmap" "Imports" 
db2$package_dependencies("ComplexHeatmap")
#>       package          dependency      dep_fields
#>  [1,] "ComplexHeatmap" "methods"       "Depends" 
#>  [2,] "ComplexHeatmap" "grid"          "Depends" 
#>  [3,] "ComplexHeatmap" "graphics"      "Depends" 
#>  [4,] "ComplexHeatmap" "stats"         "Depends" 
#>  [5,] "ComplexHeatmap" "grDevices"     "Depends" 
#>  [6,] "ComplexHeatmap" "circlize"      "Imports" 
#>  [7,] "ComplexHeatmap" "GetoptLong"    "Imports" 
#>  [8,] "ComplexHeatmap" "colorspace"    "Imports" 
#>  [9,] "ComplexHeatmap" "clue"          "Imports" 
#> [10,] "ComplexHeatmap" "RColorBrewer"  "Imports" 
#> [11,] "ComplexHeatmap" "GlobalOptions" "Imports" 
#> [12,] "ComplexHeatmap" "png"           "Imports" 
#> [13,] "ComplexHeatmap" "digest"        "Imports" 
#> [14,] "ComplexHeatmap" "IRanges"       "Imports" 
#> [15,] "ComplexHeatmap" "matrixStats"   "Imports" 
#> [16,] "ComplexHeatmap" "foreach"       "Imports" 
#> [17,] "ComplexHeatmap" "doParallel"    "Imports" 
#> [18,] "ComplexHeatmap" "codetools"     "Imports" 
db2$package_dependencies("ComplexHeatmap", recursive = TRUE)
#>       package          dependency      dep_fields 
#>  [1,] "ComplexHeatmap" "methods"       "Depends"  
#>  [2,] "ComplexHeatmap" "grid"          "Depends"  
#>  [3,] "ComplexHeatmap" "graphics"      "Depends"  
#>  [4,] "ComplexHeatmap" "stats"         "Depends"  
#>  [5,] "ComplexHeatmap" "grDevices"     "Depends"  
#>  [6,] "ComplexHeatmap" "circlize"      "Imports"  
#>  [7,] "ComplexHeatmap" "GetoptLong"    "Imports"  
#>  [8,] "ComplexHeatmap" "colorspace"    "Imports"  
#>  [9,] "ComplexHeatmap" "clue"          "Imports"  
#> [10,] "ComplexHeatmap" "RColorBrewer"  "Imports"  
#> [11,] "ComplexHeatmap" "GlobalOptions" "Imports"  
#> [12,] "ComplexHeatmap" "png"           "Imports"  
#> [13,] "ComplexHeatmap" "digest"        "Imports"  
#> [14,] "ComplexHeatmap" "IRanges"       "Imports"  
#> [15,] "ComplexHeatmap" "matrixStats"   "Imports"  
#> [16,] "ComplexHeatmap" "foreach"       "Imports"  
#> [17,] "ComplexHeatmap" "doParallel"    "Imports"  
#> [18,] "ComplexHeatmap" "codetools"     "Imports"  
#> [19,] "circlize"       "graphics"      "Depends"  
#> [20,] "circlize"       "GlobalOptions" "Imports"  
#> [21,] "circlize"       "shape"         "Imports"  
#> [22,] "circlize"       "grDevices"     "Imports"  
#> [23,] "circlize"       "utils"         "Imports"  
#> [24,] "circlize"       "stats"         "Imports"  
#> [25,] "circlize"       "colorspace"    "Imports"  
#> [26,] "circlize"       "methods"       "Imports"  
#> [27,] "circlize"       "grid"          "Imports"  
#> [28,] "GetoptLong"     "rjson"         "Imports"  
#> [29,] "GetoptLong"     "GlobalOptions" "Imports"  
#> [30,] "GetoptLong"     "methods"       "Imports"  
#> [31,] "GetoptLong"     "crayon"        "Imports"  
#> [32,] "colorspace"     "methods"       "Depends"  
#> [33,] "colorspace"     "graphics"      "Imports"  
#> [34,] "colorspace"     "grDevices"     "Imports"  
#> [35,] "colorspace"     "stats"         "Imports"  
#> [36,] "clue"           "stats"         "Imports"  
#> [37,] "clue"           "cluster"       "Imports"  
#> [38,] "clue"           "graphics"      "Imports"  
#> [39,] "clue"           "methods"       "Imports"  
#> [40,] "GlobalOptions"  "methods"       "Depends"  
#> [41,] "GlobalOptions"  "utils"         "Imports"  
#> [42,] "digest"         "utils"         "Imports"  
#> [43,] "IRanges"        "methods"       "Depends"  
#> [44,] "IRanges"        "utils"         "Depends"  
#> [45,] "IRanges"        "stats"         "Depends"  
#> [46,] "IRanges"        "BiocGenerics"  "Depends"  
#> [47,] "IRanges"        "S4Vectors"     "Depends"  
#> [48,] "IRanges"        "stats4"        "Imports"  
#> [49,] "IRanges"        "S4Vectors"     "LinkingTo"
#> [50,] "foreach"        "codetools"     "Imports"  
#> [51,] "foreach"        "utils"         "Imports"  
#> [52,] "foreach"        "iterators"     "Imports"  
#> [53,] "doParallel"     "foreach"       "Depends"  
#> [54,] "doParallel"     "iterators"     "Depends"  
#> [55,] "doParallel"     "parallel"      "Depends"  
#> [56,] "doParallel"     "utils"         "Depends"  
#> [57,] "GlobalOptions"  "methods"       "Depends"  
#> [58,] "GlobalOptions"  "utils"         "Imports"  
#> [59,] "shape"          "stats"         "Imports"  
#> [60,] "shape"          "graphics"      "Imports"  
#> [61,] "shape"          "grDevices"     "Imports"  
#> [62,] "colorspace"     "methods"       "Depends"  
#> [63,] "colorspace"     "graphics"      "Imports"  
#> [64,] "colorspace"     "grDevices"     "Imports"  
#> [65,] "colorspace"     "stats"         "Imports"  
#> [66,] "crayon"         "grDevices"     "Imports"  
#> [67,] "crayon"         "methods"       "Imports"  
#> [68,] "crayon"         "utils"         "Imports"  
#> [69,] "cluster"        "graphics"      "Imports"  
#> [70,] "cluster"        "grDevices"     "Imports"  
#> [71,] "cluster"        "stats"         "Imports"  
#> [72,] "cluster"        "utils"         "Imports"  
#> [73,] "BiocGenerics"   "methods"       "Depends"  
#> [74,] "BiocGenerics"   "utils"         "Depends"  
#> [75,] "BiocGenerics"   "graphics"      "Depends"  
#> [76,] "BiocGenerics"   "stats"         "Depends"  
#> [77,] "BiocGenerics"   "methods"       "Imports"  
#> [78,] "BiocGenerics"   "utils"         "Imports"  
#> [79,] "BiocGenerics"   "graphics"      "Imports"  
#> [80,] "BiocGenerics"   "stats"         "Imports"  
#> [81,] "S4Vectors"      "methods"       "Depends"  
#> [82,] "S4Vectors"      "utils"         "Depends"  
#> [83,] "S4Vectors"      "stats"         "Depends"  
#> [84,] "S4Vectors"      "stats4"        "Depends"  
#> [85,] "S4Vectors"      "BiocGenerics"  "Depends"  
#> [86,] "iterators"      "utils"         "Depends"  
#> [87,] "foreach"        "codetools"     "Imports"  
#> [88,] "foreach"        "utils"         "Imports"  
#> [89,] "foreach"        "iterators"     "Imports"  
#> [90,] "BiocGenerics"   "methods"       "Depends"  
#> [91,] "BiocGenerics"   "utils"         "Depends"  
#> [92,] "BiocGenerics"   "graphics"      "Depends"  
#> [93,] "BiocGenerics"   "stats"         "Depends"  
#> [94,] "BiocGenerics"   "methods"       "Imports"  
#> [95,] "BiocGenerics"   "utils"         "Imports"  
#> [96,] "BiocGenerics"   "graphics"      "Imports"  
#> [97,] "BiocGenerics"   "stats"         "Imports"  
#> [98,] "iterators"      "utils"         "Depends"  
# }