Get downstream dependency for a package

Get downstream dependency for a package

downstream_dependency(package, snapshot = TRUE)

Arguments

package

Package name.

snapshot

If it is TRUE, the package database generated on 2021-10-28 is used. If it is FALSE, the pakage database is directly retrieved from CRAN/Bioconductor.

Details

Downstream packages with relations of Depends, Imports and LinkingTo are retrieved.

Value

A data frame with all downstream packages.

Examples

# \dontrun{
downstream_dependency("ComplexHeatmap")
#>             parents                  children dep_fields heaviness
#> 2    ComplexHeatmap                  AMARETTO    Depends         5
#> 21   ComplexHeatmap           EnrichedHeatmap    Depends        11
#> 30   ComplexHeatmap InteractiveComplexHeatmap    Depends        12
#> 61   ComplexHeatmap               countToFPKM    Depends        19
#> 47   ComplexHeatmap                    recoup    Depends         7
#> 3    ComplexHeatmap                    ASURAT    Imports        13
#> 5    ComplexHeatmap                   BioNERO    Imports         7
#> 4    ComplexHeatmap                BiocOncoTK    Imports        10
#> 7    ComplexHeatmap           BloodGen3Module    Imports         9
#> 8    ComplexHeatmap                  CATALYST    Imports         6
#> 11   ComplexHeatmap                     COCOA    Imports        11
#> 13   ComplexHeatmap                     COTAN    Imports        14
#> 10   ComplexHeatmap                      CeTF    Imports         8
#> 15   ComplexHeatmap          DEComplexDisease    Imports        11
#> 16   ComplexHeatmap                 DEGreport    Imports         7
#> 17   ComplexHeatmap                       DEP    Imports         4
#> 20   ComplexHeatmap                     ELMER    Imports         3
#> 23   ComplexHeatmap                    FLAMES    Imports         6
#> 28   ComplexHeatmap                    GRaNIE    Imports         8
#> 25   ComplexHeatmap                 GeneTonic    Imports         8
#> 26   ComplexHeatmap     GenomicSuperSignature    Imports        13
#> 31   ComplexHeatmap               InterCellar    Imports         9
#> 33   ComplexHeatmap              LineagePulse    Imports         9
#> 63   ComplexHeatmap                   MKomics    Imports        18
#> 37   ComplexHeatmap                      MOMA    Imports         9
#> 41   ComplexHeatmap                 MWASTools    Imports        12
#> 34   ComplexHeatmap               MatrixQCvis    Imports        12
#> 35   ComplexHeatmap                    MesKit    Imports        10
#> 62   ComplexHeatmap                  MitoHEAR    Imports        10
#> 64   ComplexHeatmap                     PALMO    Imports         7
#> 45   ComplexHeatmap                      POMA    Imports         8
#> 42   ComplexHeatmap                 PathoStat    Imports         7
#> 43   ComplexHeatmap                   PeacoQC    Imports        10
#> 48   ComplexHeatmap                     RLSeq    Imports         5
#> 67   ComplexHeatmap                  RNAseqQC    Imports         7
#> 68   ComplexHeatmap                       RVA    Imports         8
#> 54   ComplexHeatmap                    SPONGE    Imports         9
#> 55   ComplexHeatmap       TBSignatureProfiler    Imports        11
#> 56   ComplexHeatmap                      Xeva    Imports         8
#> 57   ComplexHeatmap                     YAPSA    Imports         2
#> 1    ComplexHeatmap                   airpart    Imports         7
#> 58   ComplexHeatmap                    armada    Imports        11
#> 6    ComplexHeatmap               blacksheepr    Imports         8
#> 59   ComplexHeatmap              bulkAnalyseR    Imports         5
#> 9    ComplexHeatmap                     celda    Imports         4
#> 12   ComplexHeatmap                      cola    Imports         4
#> 60   ComplexHeatmap                     conos    Imports        18
#> 14   ComplexHeatmap                cytoKernel    Imports        11
#> 19   ComplexHeatmap                   diffUTR    Imports        10
#> 18   ComplexHeatmap                   diffcyt    Imports         6
#> 22   ComplexHeatmap                     fCCAC    Imports        10
#> 24   ComplexHeatmap              gCrisprTools    Imports        13
#> 27   ComplexHeatmap                    gmoviz    Imports         8
#> 29   ComplexHeatmap                    hermes    Imports         8
#> 32   ComplexHeatmap                      iSEE    Imports        11
#> 36   ComplexHeatmap          microbiomeMarker    Imports         7
#> 38   ComplexHeatmap                  monaLisa    Imports         5
#> 39   ComplexHeatmap                    muscat    Imports         8
#> 40   ComplexHeatmap                  musicatk    Imports         6
#> 44   ComplexHeatmap                  pipeComp    Imports         6
#> 65   ComplexHeatmap                   pkgndep    Imports        21
#> 46   ComplexHeatmap               profileplyr    Imports         0
#> 66   ComplexHeatmap                   rKOMICS    Imports        14
#> 69   ComplexHeatmap                     scITD    Imports         4
#> 49   ComplexHeatmap                     sechm    Imports         9
#> 50   ComplexHeatmap                 segmenter    Imports        10
#> 51   ComplexHeatmap        simplifyEnrichment    Imports         7
#> 52   ComplexHeatmap              singleCellTK    Imports         0
#> 53   ComplexHeatmap                   sparrow    Imports         8
#> 70   ComplexHeatmap               tidyHeatmap    Imports        16
#> 71   ComplexHeatmap                visxhclust    Imports        15
#> 72   ComplexHeatmap                    wilson    Imports         4
#> 110 EnrichedHeatmap                extraChIPs    Imports        10
#> 210 EnrichedHeatmap               profileplyr    Imports         2
#> 111         BioNERO                 cageminer    Imports        77
#> 112       DEGreport                   isomiRs    Imports        26
#> 113           ELMER           TCGAbiolinksGUI    Imports        71
#> 114       MWASTools              MetaboSignal    Imports        44
#> 115          SPONGE                miRspongeR    Imports       136
#> 116           celda              singleCellTK    Imports        10
#> 117         diffcyt                   censcyt    Depends       113
#> 211         diffcyt                CyTOFpower    Imports        83
#> 310         diffcyt                   treekoR    Imports       114
#> 118            iSEE                     iSEEu    Depends        21
#> 119           sechm                   SEtools    Imports        21
#> 120      miRspongeR                     miRSM    Imports       147
# }