Load all package dependency statistics

Load all package dependency statistics

load_pkg_stat_snapshot()

Details

It loads the package dependency analysis for all CRAN/Biocondutor packages done on 2021-10-28.

Value

A data frame of various columns.

Examples

# \dontrun{
df = load_pkg_stat_snapshot()
head(df)
#>                 package                                              repository
#> a4                   a4 https://bioconductor.org/packages/3.14/bioc/src/contrib
#> a4Base           a4Base https://bioconductor.org/packages/3.14/bioc/src/contrib
#> a4Classif     a4Classif https://bioconductor.org/packages/3.14/bioc/src/contrib
#> a4Core           a4Core https://bioconductor.org/packages/3.14/bioc/src/contrib
#> a4Preproc     a4Preproc https://bioconductor.org/packages/3.14/bioc/src/contrib
#> a4Reporting a4Reporting https://bioconductor.org/packages/3.14/bioc/src/contrib
#>             n_by_strong n_by_all n_parents gini_index
#> a4                   81      161         5  0.7803030
#> a4Base               72      153        13  0.2276423
#> a4Classif            30       84        11  0.2914286
#> a4Core               17       36         4  0.5555556
#> a4Preproc             6       62         2  0.2000000
#> a4Reporting           4       24         2  0.2500000
#>             max_heaviness_from_parents adjusted_max_heaviness_from_parents
#> a4                                  48                          21.8181818
#> a4Base                               4                           2.2337662
#> a4Classif                            6                           3.1948052
#> a4Core                              11                           4.8571429
#> a4Preproc                            1                           0.4155844
#> a4Reporting                          3                           1.2467532
#>                                                                                                                                                                                                      max_heaviness_parent_info
#> a4                  Parent package &lsquo;a4Base&rsquo; contibutes the highest heaviness on &lsquo;a4&rsquo;. It is listed in &lsquo;Depends&rsquo but no object from parent is imported to the namespace of &lsquo;a4&rsquo;.
#> a4Base                                            Parent package &lsquo;genefilter&rsquo; contibutes the highest heaviness on &lsquo;a4Base&rsquo;. 5 functions/objects are imported to the namespace of &lsquo;a4Base&rsquo;.
#> a4Classif                                         Parent package &lsquo;ROCR&rsquo; contibutes the highest heaviness on &lsquo;a4Classif&rsquo;. 3 functions/objects are imported to the namespace of &lsquo;a4Classif&rsquo;.
#> a4Core      Parent package &lsquo;glmnet&rsquo; contibutes the highest heaviness on &lsquo;a4Core&rsquo;. The whole set of functions/methods/classes from parent package is imported to the namespace of &lsquo;a4Core&rsquo;.
#> a4Preproc                                      Parent package &lsquo;Biobase&rsquo; contibutes the highest heaviness on &lsquo;a4Preproc&rsquo;. 7 functions/objects are imported to the namespace of &lsquo;a4Preproc&rsquo;.
#> a4Reporting                                 Parent package &lsquo;xtable&rsquo; contibutes the highest heaviness on &lsquo;a4Reporting&rsquo;. 1 functions/objects are imported to the namespace of &lsquo;a4Reporting&rsquo;.
#>             total_heaviness_from_parents adjusted_total_heaviness_from_parents
#> a4                                    56                            31.9089614
#> a4Base                                15                             9.8198051
#> a4Classif                             13                             8.2485733
#> a4Core                                13                             7.2577768
#> a4Preproc                              1                             0.5345225
#> a4Reporting                            4                             2.1380899
#>             reducible heaviness_on_children n_children heaviness_on_downstream
#> a4              FALSE             0.0000000          0                       0
#> a4Base           TRUE            48.0000000          1                      48
#> a4Classif        TRUE             7.0000000          1                       7
#> a4Core          FALSE             1.7500000          4                       2
#> a4Preproc        TRUE             0.6666667          3                       1
#> a4Reporting      TRUE             1.0000000          1                       1
#>             n_downstream hv_downstream_values adjusted_heaviness_on_children
#> a4                     0                                          0.00000000
#> a4Base                 1                   48                     4.36363636
#> a4Classif              1                    7                     0.63636364
#> a4Core                 4           1, 1, 1, 5                     0.50000000
#> a4Preproc              3              1, 1, 1                     0.15384615
#> a4Reporting            1                    1                     0.09090909
#>             adjusted_heaviness_on_downstream
#> a4                                 0.0000000
#> a4Base                             3.0000000
#> a4Classif                          0.4375000
#> a4Core                             0.4210526
#> a4Preproc                          0.1666667
#> a4Reporting                        0.0625000
#>             heaviness_on_downstream_no_children n_downstream_no_children
#> a4                                            0                        0
#> a4Base                                        0                        0
#> a4Classif                                     0                        0
#> a4Core                                        0                        0
#> a4Preproc                                     0                        0
#> a4Reporting                                   0                        0
#>             hv_downstream_no_children_values
#> a4                                          
#> a4Base                                      
#> a4Classif                                   
#> a4Core                                      
#> a4Preproc                                   
#> a4Reporting                                 
#>             adjusted_heaviness_on_downstream_no_children
#> a4                                                     0
#> a4Base                                                 0
#> a4Classif                                              0
#> a4Core                                                 0
#> a4Preproc                                              0
#> a4Reporting                                            0
# }