Load all package dependency statistics

Load all package dependency statistics

load_pkg_stat_snapshot()

Details

It is calculated based on a specific CRAN/Bioconductor snapshot. The version is set via pkgndep_opt$heaviness_db_version.

Value

A data frame.

Examples

# \dontrun{
df = load_pkg_stat_snapshot()
head(df)
#>                 package version repository n_by_strong n_by_all n_parents
#> A3                   A3   1.0.0       CRAN           5       13         2
#> AATtools       AATtools   0.0.2       CRAN          26       26         4
#> ABACUS           ABACUS   1.0.0       CRAN          60       69         2
#> ABC.RAP         ABC.RAP   0.9.0       CRAN           3       31         3
#> ABCanalysis ABCanalysis   1.2.1       CRAN           5        5         1
#> ABCoptim       ABCoptim  0.15.0       CRAN           5       48         4
#>             max_rel_heaviness_from_parents max_heaviness_from_parents
#> A3                                1.250000                          3
#> AATtools                          2.117647                         19
#> ABACUS                            1.521739                         24
#> ABC.RAP                           1.083333                          1
#> ABCanalysis                       1.500000                          5
#> ABCoptim                          1.153846                          2
#>             adjusted_max_heaviness_from_parents heaviest_parent
#> A3                                    1.1927711          xtable
#> AATtools                              8.0120482           dplyr
#> ABACUS                                9.5421687           shiny
#> ABC.RAP                               0.4096386        graphics
#> ABCanalysis                           1.9277108         plotrix
#> ABCoptim                              0.8433735            Rcpp
#>                                                                                                                                                                                                          max_heaviness_parent_info
#> A3                      Parent package &lsquo;xtable&rsquo; contibutes the highest heaviness on &lsquo;A3&rsquo;. The whole set of functions/methods/classes from parent package is imported to the namespace of &lsquo;A3&rsquo;.
#> AATtools                                               Parent package &lsquo;dplyr&rsquo; contibutes the highest heaviness on &lsquo;AATtools&rsquo;. 8 functions/objects are imported to the namespace of &lsquo;AATtools&rsquo;.
#> ABACUS           Parent package &lsquo;shiny&rsquo; contibutes the highest heaviness on &lsquo;ABACUS&rsquo;. The whole set of functions/methods/classes from parent package is imported to the namespace of &lsquo;ABACUS&rsquo;.
#> ABC.RAP     Parent package &lsquo;graphics&rsquo; contibutes the highest heaviness on &lsquo;ABC.RAP&rsquo;. The whole set of functions/methods/classes from parent package is imported to the namespace of &lsquo;ABC.RAP&rsquo;.
#> ABCanalysis                                    Parent package &lsquo;plotrix&rsquo; contibutes the highest heaviness on &lsquo;ABCanalysis&rsquo;. 1 functions/objects are imported to the namespace of &lsquo;ABCanalysis&rsquo;.
#> ABCoptim                                                Parent package &lsquo;Rcpp&rsquo; contibutes the highest heaviness on &lsquo;ABCoptim&rsquo;. 1 functions/objects are imported to the namespace of &lsquo;ABCoptim&rsquo;.
#>             total_heaviness_from_parents reducible
#> A3                                     5     FALSE
#> AATtools                              21      TRUE
#> ABACUS                                47     FALSE
#> ABC.RAP                                3     FALSE
#> ABCanalysis                            5      TRUE
#> ABCoptim                               4      TRUE
#>             max_co_heaviness_from_parents max_co_heaviness_parents_pair
#> A3                                      0                              
#> AATtools                                3            doParallel,foreach
#> ABACUS                                 13                 ggplot2,shiny
#> ABC.RAP                                 0                              
#> ABCanalysis                             0                              
#> ABCoptim                                1                    Rcpp,utils
#>             max_co_heaviness_parents_pair_type heaviness_on_children n_children
#> A3                                                                 0          0
#> AATtools                          parent-child                     0          0
#> ABACUS                            parent-child                     0          0
#> ABC.RAP                                                            0          0
#> ABCanalysis                                                        2          1
#> ABCoptim                          parent-child                     0          0
#>             heaviness_on_downstream n_downstream hv_downstream_values
#> A3                                0            0                     
#> AATtools                          0            0                     
#> ABACUS                            0            0                     
#> ABC.RAP                           0            0                     
#> ABCanalysis                       2            1                    2
#> ABCoptim                          0            0                     
#>             adjusted_heaviness_on_children adjusted_heaviness_on_downstream
#> A3                               0.0000000                        0.0000000
#> AATtools                         0.0000000                        0.0000000
#> ABACUS                           0.0000000                        0.0000000
#> ABC.RAP                          0.0000000                        0.0000000
#> ABCanalysis                      0.1666667                        0.1333333
#> ABCoptim                         0.0000000                        0.0000000
#>             heaviness_on_indirect_downstream n_indirect_downstream
#> A3                                         0                     0
#> AATtools                                   0                     0
#> ABACUS                                     0                     0
#> ABC.RAP                                    0                     0
#> ABCanalysis                                0                     0
#> ABCoptim                                   0                     0
#>             hv_indirect_downstream_values
#> A3                                       
#> AATtools                                 
#> ABACUS                                   
#> ABC.RAP                                  
#> ABCanalysis                              
#> ABCoptim                                 
#>             adjusted_heaviness_on_indirect_downstream
#> A3                                                  0
#> AATtools                                            0
#> ABACUS                                              0
#> ABC.RAP                                             0
#> ABCanalysis                                         0
#> ABCoptim                                            0
# }