Loaded packages

Loaded packages

loaded_packages(pkg, verbose = TRUE)

Arguments

pkg

A package name.

verbose

Whether to print messages.

Details

It loads pkg into a new R session and collects which other packages are loaded by parsing the output from sessionInfo.

Value

A data frame.

Examples

loaded_packages("ComplexHeatmap")
#> Loading ComplexHeatmap to a new R session... 31 namespaces loaded.
#>              pkg       type loading_time
#> 1           grid   basePkgs        1.331
#> 2          stats   basePkgs        1.331
#> 3       graphics   basePkgs        1.331
#> 4      grDevices   basePkgs        1.331
#> 5          utils   basePkgs        1.331
#> 6       datasets   basePkgs        1.331
#> 7        methods   basePkgs        1.331
#> 8           base   basePkgs        1.331
#> 9       circlize loadedOnly        1.331
#> 10   matrixStats loadedOnly        1.331
#> 11     codetools loadedOnly        1.331
#> 12       IRanges loadedOnly        1.331
#> 13           png loadedOnly        1.331
#> 14        digest loadedOnly        1.331
#> 15       foreach loadedOnly        1.331
#> 16        crayon loadedOnly        1.331
#> 17        stats4 loadedOnly        1.331
#> 18 GlobalOptions loadedOnly        1.331
#> 19    doParallel loadedOnly        1.331
#> 20     S4Vectors loadedOnly        1.331
#> 21    GetoptLong loadedOnly        1.331
#> 22  RColorBrewer loadedOnly        1.331
#> 23         rjson loadedOnly        1.331
#> 24     iterators loadedOnly        1.331
#> 25      parallel loadedOnly        1.331
#> 26      compiler loadedOnly        1.331
#> 27          clue loadedOnly        1.331
#> 28    colorspace loadedOnly        1.331
#> 29  BiocGenerics loadedOnly        1.331
#> 30       cluster loadedOnly        1.331
#> 31         shape loadedOnly        1.331