Format the package database

Format the package database

reformat_db(db)

Arguments

db

A data frame returned from available.packages or installed.packages.

Details

It reformats the data frame of the package database into a pkg_db class object.

Value

A pkg_db class object. There are the following methods:

pkg_db$get_meta(package,field=NULL)

field can take values in "Package", "Version" and "Repository".

pkg_db$get_dependency_table(package)

Get the dependency table.

pkg_db$get_rev_dependency_table(package)

Get the reverse dependency table.

pkg_db$package_dependencies(package,recursive=FALSE,reverse=FALSE,which="strong",simplify=FALSE)

All the arguments are the same as in package_dependencies. Argument simplify controls whether to return a data frame or a simplied vector.

Examples

# \dontrun{
db = available.packages()
db2 = reformat_db(db)
#> prepare dependency table...
#> prepare reverse dependency table...

# a pkg_db object generated on 2021-10-28 can be loaded by load_pkg_db()
db2 = load_pkg_db(snapshot = TRUE)
db2
#> A package database of 21741 packages.
#>   - 0 local / 21741 remote packages.
#>   - 3416 Bioconductor / 18325 CRAN / 0 other packages.
db2$get_meta("ComplexHeatmap")
#>                       Package Version Local
#> ComplexHeatmap ComplexHeatmap  2.10.0    no
#>                                                             Repository
#> ComplexHeatmap https://bioconductor.org/packages/3.14/bioc/src/contrib
db2$get_dependency_table("ComplexHeatmap")
#>       package          dependency        dep_fields
#>  [1,] "ComplexHeatmap" "methods"         "Depends" 
#>  [2,] "ComplexHeatmap" "grid"            "Depends" 
#>  [3,] "ComplexHeatmap" "graphics"        "Depends" 
#>  [4,] "ComplexHeatmap" "stats"           "Depends" 
#>  [5,] "ComplexHeatmap" "grDevices"       "Depends" 
#>  [6,] "ComplexHeatmap" "circlize"        "Imports" 
#>  [7,] "ComplexHeatmap" "GetoptLong"      "Imports" 
#>  [8,] "ComplexHeatmap" "colorspace"      "Imports" 
#>  [9,] "ComplexHeatmap" "clue"            "Imports" 
#> [10,] "ComplexHeatmap" "RColorBrewer"    "Imports" 
#> [11,] "ComplexHeatmap" "GlobalOptions"   "Imports" 
#> [12,] "ComplexHeatmap" "png"             "Imports" 
#> [13,] "ComplexHeatmap" "digest"          "Imports" 
#> [14,] "ComplexHeatmap" "IRanges"         "Imports" 
#> [15,] "ComplexHeatmap" "matrixStats"     "Imports" 
#> [16,] "ComplexHeatmap" "foreach"         "Imports" 
#> [17,] "ComplexHeatmap" "doParallel"      "Imports" 
#> [18,] "ComplexHeatmap" "testthat"        "Suggests"
#> [19,] "ComplexHeatmap" "knitr"           "Suggests"
#> [20,] "ComplexHeatmap" "markdown"        "Suggests"
#> [21,] "ComplexHeatmap" "dendsort"        "Suggests"
#> [22,] "ComplexHeatmap" "jpeg"            "Suggests"
#> [23,] "ComplexHeatmap" "tiff"            "Suggests"
#> [24,] "ComplexHeatmap" "fastcluster"     "Suggests"
#> [25,] "ComplexHeatmap" "EnrichedHeatmap" "Suggests"
#> [26,] "ComplexHeatmap" "dendextend"      "Suggests"
#> [27,] "ComplexHeatmap" "grImport"        "Suggests"
#> [28,] "ComplexHeatmap" "grImport2"       "Suggests"
#> [29,] "ComplexHeatmap" "glue"            "Suggests"
#> [30,] "ComplexHeatmap" "GenomicRanges"   "Suggests"
#> [31,] "ComplexHeatmap" "gridtext"        "Suggests"
#> [32,] "ComplexHeatmap" "pheatmap"        "Suggests"
#> [33,] "ComplexHeatmap" "gridGraphics"    "Suggests"
#> [34,] "ComplexHeatmap" "gplots"          "Suggests"
#> [35,] "ComplexHeatmap" "rmarkdown"       "Suggests"
#> [36,] "ComplexHeatmap" "Cairo"           "Suggests"
db2$get_rev_dependency_table("ComplexHeatmap")
#>       package                     dependency       dep_fields
#>  [1,] "airpart"                   "ComplexHeatmap" "Imports" 
#>  [2,] "AMARETTO"                  "ComplexHeatmap" "Depends" 
#>  [3,] "artMS"                     "ComplexHeatmap" "Suggests"
#>  [4,] "bambu"                     "ComplexHeatmap" "Suggests"
#>  [5,] "BindingSiteFinder"         "ComplexHeatmap" "Suggests"
#>  [6,] "BiocOncoTK"                "ComplexHeatmap" "Imports" 
#>  [7,] "BioNERO"                   "ComplexHeatmap" "Imports" 
#>  [8,] "blacksheepr"               "ComplexHeatmap" "Imports" 
#>  [9,] "BloodGen3Module"           "ComplexHeatmap" "Imports" 
#> [10,] "BrainSABER"                "ComplexHeatmap" "Suggests"
#> [11,] "CATALYST"                  "ComplexHeatmap" "Imports" 
#> [12,] "celda"                     "ComplexHeatmap" "Imports" 
#> [13,] "CeTF"                      "ComplexHeatmap" "Imports" 
#> [14,] "clustifyr"                 "ComplexHeatmap" "Suggests"
#> [15,] "CNVRanger"                 "ComplexHeatmap" "Suggests"
#> [16,] "COCOA"                     "ComplexHeatmap" "Imports" 
#> [17,] "cola"                      "ComplexHeatmap" "Imports" 
#> [18,] "cytoKernel"                "ComplexHeatmap" "Imports" 
#> [19,] "DEComplexDisease"          "ComplexHeatmap" "Imports" 
#> [20,] "DEGreport"                 "ComplexHeatmap" "Imports" 
#> [21,] "DEP"                       "ComplexHeatmap" "Imports" 
#> [22,] "diffcyt"                   "ComplexHeatmap" "Imports" 
#> [23,] "diffUTR"                   "ComplexHeatmap" "Imports" 
#> [24,] "dittoSeq"                  "ComplexHeatmap" "Suggests"
#> [25,] "ELMER"                     "ComplexHeatmap" "Imports" 
#> [26,] "EnrichedHeatmap"           "ComplexHeatmap" "Depends" 
#> [27,] "EnrichmentBrowser"         "ComplexHeatmap" "Suggests"
#> [28,] "fCCAC"                     "ComplexHeatmap" "Imports" 
#> [29,] "gCrisprTools"              "ComplexHeatmap" "Imports" 
#> [30,] "GeneTonic"                 "ComplexHeatmap" "Imports" 
#> [31,] "GenomicSuperSignature"     "ComplexHeatmap" "Imports" 
#> [32,] "gmoviz"                    "ComplexHeatmap" "Imports" 
#> [33,] "gtrellis"                  "ComplexHeatmap" "Suggests"
#> [34,] "HilbertCurve"              "ComplexHeatmap" "Suggests"
#> [35,] "InteractiveComplexHeatmap" "ComplexHeatmap" "Imports" 
#> [36,] "InterCellar"               "ComplexHeatmap" "Imports" 
#> [37,] "iSEE"                      "ComplexHeatmap" "Imports" 
#> [38,] "LineagePulse"              "ComplexHeatmap" "Imports" 
#> [39,] "MatrixQCvis"               "ComplexHeatmap" "Imports" 
#> [40,] "MesKit"                    "ComplexHeatmap" "Imports" 
#> [41,] "microbiomeMarker"          "ComplexHeatmap" "Imports" 
#> [42,] "MOMA"                      "ComplexHeatmap" "Imports" 
#> [43,] "monaLisa"                  "ComplexHeatmap" "Imports" 
#> [44,] "msImpute"                  "ComplexHeatmap" "Suggests"
#> [45,] "muscat"                    "ComplexHeatmap" "Imports" 
#> [46,] "musicatk"                  "ComplexHeatmap" "Imports" 
#> [47,] "MWASTools"                 "ComplexHeatmap" "Imports" 
#> [48,] "PathoStat"                 "ComplexHeatmap" "Imports" 
#> [49,] "PeacoQC"                   "ComplexHeatmap" "Imports" 
#> [50,] "pipeComp"                  "ComplexHeatmap" "Imports" 
#> [51,] "POMA"                      "ComplexHeatmap" "Imports" 
#> [52,] "profileplyr"               "ComplexHeatmap" "Imports" 
#> [53,] "projectR"                  "ComplexHeatmap" "Suggests"
#> [54,] "QFeatures"                 "ComplexHeatmap" "Suggests"
#> [55,] "recoup"                    "ComplexHeatmap" "Depends" 
#> [56,] "RLSeq"                     "ComplexHeatmap" "Imports" 
#> [57,] "scDblFinder"               "ComplexHeatmap" "Suggests"
#> [58,] "sechm"                     "ComplexHeatmap" "Imports" 
#> [59,] "segmenter"                 "ComplexHeatmap" "Imports" 
#> [60,] "SEtools"                   "ComplexHeatmap" "Imports" 
#> [61,] "simplifyEnrichment"        "ComplexHeatmap" "Imports" 
#> [62,] "singleCellTK"              "ComplexHeatmap" "Imports" 
#> [63,] "sparrow"                   "ComplexHeatmap" "Imports" 
#> [64,] "spiky"                     "ComplexHeatmap" "Suggests"
#> [65,] "TBSignatureProfiler"       "ComplexHeatmap" "Imports" 
#> [66,] "TCGAbiolinks"              "ComplexHeatmap" "Suggests"
#> [67,] "TCGAutils"                 "ComplexHeatmap" "Suggests"
#> [68,] "TimeSeriesExperiment"      "ComplexHeatmap" "Suggests"
#> [69,] "weitrix"                   "ComplexHeatmap" "Suggests"
#> [70,] "Xeva"                      "ComplexHeatmap" "Imports" 
#> [71,] "YAPSA"                     "ComplexHeatmap" "Imports" 
#> [72,] "NanoporeRNASeq"            "ComplexHeatmap" "Suggests"
#> [73,] "TCGAWorkflow"              "ComplexHeatmap" "Imports" 
#> [74,] "armada"                    "ComplexHeatmap" "Imports" 
#> [75,] "circlize"                  "ComplexHeatmap" "Suggests"
#> [76,] "CNVScope"                  "ComplexHeatmap" "Suggests"
#> [77,] "conos"                     "ComplexHeatmap" "Imports" 
#> [78,] "countToFPKM"               "ComplexHeatmap" "Depends" 
#> [79,] "eclust"                    "ComplexHeatmap" "Suggests"
#> [80,] "i2dash"                    "ComplexHeatmap" "Suggests"
#> [81,] "IOHanalyzer"               "ComplexHeatmap" "Suggests"
#> [82,] "MKomics"                   "ComplexHeatmap" "Imports" 
#> [83,] "MOSS"                      "ComplexHeatmap" "Suggests"
#> [84,] "multipanelfigure"          "ComplexHeatmap" "Suggests"
#> [85,] "pkgndep"                   "ComplexHeatmap" "Imports" 
#> [86,] "rKOMICS"                   "ComplexHeatmap" "Imports" 
#> [87,] "RVA"                       "ComplexHeatmap" "Imports" 
#> [88,] "sigQC"                     "ComplexHeatmap" "Imports" 
#> [89,] "spiralize"                 "ComplexHeatmap" "Suggests"
#> [90,] "tidyHeatmap"               "ComplexHeatmap" "Imports" 
#> [91,] "visxhclust"                "ComplexHeatmap" "Imports" 
#> [92,] "wilson"                    "ComplexHeatmap" "Imports" 
db2$package_dependencies("ComplexHeatmap")
#>       package          dependency      dep_fields
#>  [1,] "ComplexHeatmap" "methods"       "Depends" 
#>  [2,] "ComplexHeatmap" "grid"          "Depends" 
#>  [3,] "ComplexHeatmap" "graphics"      "Depends" 
#>  [4,] "ComplexHeatmap" "stats"         "Depends" 
#>  [5,] "ComplexHeatmap" "grDevices"     "Depends" 
#>  [6,] "ComplexHeatmap" "circlize"      "Imports" 
#>  [7,] "ComplexHeatmap" "GetoptLong"    "Imports" 
#>  [8,] "ComplexHeatmap" "colorspace"    "Imports" 
#>  [9,] "ComplexHeatmap" "clue"          "Imports" 
#> [10,] "ComplexHeatmap" "RColorBrewer"  "Imports" 
#> [11,] "ComplexHeatmap" "GlobalOptions" "Imports" 
#> [12,] "ComplexHeatmap" "png"           "Imports" 
#> [13,] "ComplexHeatmap" "digest"        "Imports" 
#> [14,] "ComplexHeatmap" "IRanges"       "Imports" 
#> [15,] "ComplexHeatmap" "matrixStats"   "Imports" 
#> [16,] "ComplexHeatmap" "foreach"       "Imports" 
#> [17,] "ComplexHeatmap" "doParallel"    "Imports" 
db2$package_dependencies("ComplexHeatmap", recursive = TRUE)
#>       package          dependency      dep_fields 
#>  [1,] "ComplexHeatmap" "methods"       "Depends"  
#>  [2,] "ComplexHeatmap" "grid"          "Depends"  
#>  [3,] "ComplexHeatmap" "graphics"      "Depends"  
#>  [4,] "ComplexHeatmap" "stats"         "Depends"  
#>  [5,] "ComplexHeatmap" "grDevices"     "Depends"  
#>  [6,] "ComplexHeatmap" "circlize"      "Imports"  
#>  [7,] "ComplexHeatmap" "GetoptLong"    "Imports"  
#>  [8,] "ComplexHeatmap" "colorspace"    "Imports"  
#>  [9,] "ComplexHeatmap" "clue"          "Imports"  
#> [10,] "ComplexHeatmap" "RColorBrewer"  "Imports"  
#> [11,] "ComplexHeatmap" "GlobalOptions" "Imports"  
#> [12,] "ComplexHeatmap" "png"           "Imports"  
#> [13,] "ComplexHeatmap" "digest"        "Imports"  
#> [14,] "ComplexHeatmap" "IRanges"       "Imports"  
#> [15,] "ComplexHeatmap" "matrixStats"   "Imports"  
#> [16,] "ComplexHeatmap" "foreach"       "Imports"  
#> [17,] "ComplexHeatmap" "doParallel"    "Imports"  
#> [18,] "circlize"       "graphics"      "Depends"  
#> [19,] "circlize"       "GlobalOptions" "Imports"  
#> [20,] "circlize"       "shape"         "Imports"  
#> [21,] "circlize"       "grDevices"     "Imports"  
#> [22,] "circlize"       "utils"         "Imports"  
#> [23,] "circlize"       "stats"         "Imports"  
#> [24,] "circlize"       "colorspace"    "Imports"  
#> [25,] "circlize"       "methods"       "Imports"  
#> [26,] "circlize"       "grid"          "Imports"  
#> [27,] "GetoptLong"     "rjson"         "Imports"  
#> [28,] "GetoptLong"     "GlobalOptions" "Imports"  
#> [29,] "GetoptLong"     "methods"       "Imports"  
#> [30,] "GetoptLong"     "crayon"        "Imports"  
#> [31,] "colorspace"     "methods"       "Depends"  
#> [32,] "colorspace"     "graphics"      "Imports"  
#> [33,] "colorspace"     "grDevices"     "Imports"  
#> [34,] "colorspace"     "stats"         "Imports"  
#> [35,] "clue"           "stats"         "Imports"  
#> [36,] "clue"           "cluster"       "Imports"  
#> [37,] "clue"           "graphics"      "Imports"  
#> [38,] "clue"           "methods"       "Imports"  
#> [39,] "GlobalOptions"  "methods"       "Depends"  
#> [40,] "GlobalOptions"  "utils"         "Imports"  
#> [41,] "digest"         "utils"         "Imports"  
#> [42,] "IRanges"        "methods"       "Depends"  
#> [43,] "IRanges"        "utils"         "Depends"  
#> [44,] "IRanges"        "stats"         "Depends"  
#> [45,] "IRanges"        "BiocGenerics"  "Depends"  
#> [46,] "IRanges"        "S4Vectors"     "Depends"  
#> [47,] "IRanges"        "stats4"        "Imports"  
#> [48,] "IRanges"        "S4Vectors"     "LinkingTo"
#> [49,] "foreach"        "codetools"     "Imports"  
#> [50,] "foreach"        "utils"         "Imports"  
#> [51,] "foreach"        "iterators"     "Imports"  
#> [52,] "doParallel"     "foreach"       "Depends"  
#> [53,] "doParallel"     "iterators"     "Depends"  
#> [54,] "doParallel"     "parallel"      "Depends"  
#> [55,] "doParallel"     "utils"         "Depends"  
#> [56,] "GlobalOptions"  "methods"       "Depends"  
#> [57,] "GlobalOptions"  "utils"         "Imports"  
#> [58,] "shape"          "stats"         "Imports"  
#> [59,] "shape"          "graphics"      "Imports"  
#> [60,] "shape"          "grDevices"     "Imports"  
#> [61,] "colorspace"     "methods"       "Depends"  
#> [62,] "colorspace"     "graphics"      "Imports"  
#> [63,] "colorspace"     "grDevices"     "Imports"  
#> [64,] "colorspace"     "stats"         "Imports"  
#> [65,] "crayon"         "grDevices"     "Imports"  
#> [66,] "crayon"         "methods"       "Imports"  
#> [67,] "crayon"         "utils"         "Imports"  
#> [68,] "cluster"        "graphics"      "Imports"  
#> [69,] "cluster"        "grDevices"     "Imports"  
#> [70,] "cluster"        "stats"         "Imports"  
#> [71,] "cluster"        "utils"         "Imports"  
#> [72,] "BiocGenerics"   "methods"       "Depends"  
#> [73,] "BiocGenerics"   "utils"         "Depends"  
#> [74,] "BiocGenerics"   "graphics"      "Depends"  
#> [75,] "BiocGenerics"   "stats"         "Depends"  
#> [76,] "BiocGenerics"   "methods"       "Imports"  
#> [77,] "BiocGenerics"   "utils"         "Imports"  
#> [78,] "BiocGenerics"   "graphics"      "Imports"  
#> [79,] "BiocGenerics"   "stats"         "Imports"  
#> [80,] "S4Vectors"      "methods"       "Depends"  
#> [81,] "S4Vectors"      "utils"         "Depends"  
#> [82,] "S4Vectors"      "stats"         "Depends"  
#> [83,] "S4Vectors"      "stats4"        "Depends"  
#> [84,] "S4Vectors"      "BiocGenerics"  "Depends"  
#> [85,] "iterators"      "utils"         "Depends"  
#> [86,] "foreach"        "codetools"     "Imports"  
#> [87,] "foreach"        "utils"         "Imports"  
#> [88,] "foreach"        "iterators"     "Imports"  
#> [89,] "BiocGenerics"   "methods"       "Depends"  
#> [90,] "BiocGenerics"   "utils"         "Depends"  
#> [91,] "BiocGenerics"   "graphics"      "Depends"  
#> [92,] "BiocGenerics"   "stats"         "Depends"  
#> [93,] "BiocGenerics"   "methods"       "Imports"  
#> [94,] "BiocGenerics"   "utils"         "Imports"  
#> [95,] "BiocGenerics"   "graphics"      "Imports"  
#> [96,] "BiocGenerics"   "stats"         "Imports"  
#> [97,] "iterators"      "utils"         "Depends"  
# }